+ cards

+++ Classes/Transactions.cs
@@ -12,13 +12,13 @@ namespace Bank.Classes
         public OperationType Operation { get; set; }
         public DateTime Timestamp { get; set; }
         public bool IsSuccessful { get; set; }
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
         public enum OperationType { Снятие, Пополнение, Перевод }
         public string GetterAccountNumber { get; set; }
         public string GetterAccountName { get; set; }
         public string SenderAccountName { get; set; }
 
-        public Transactions(string accountNumber, OperationType operation, DateTime timestamp, bool isSuccessful, decimal amount, string getterAccountNumber, string senderAccountName, string getterAccountName)
+        public Transactions(string accountNumber, OperationType operation, DateTime timestamp, bool isSuccessful, double amount, string getterAccountNumber, string senderAccountName, string getterAccountName)
         {
             AccountNumber = accountNumber;
             Operation = operation;
@@ -29,6 +29,7 @@ namespace Bank.Classes
             GetterAccountName = getterAccountName;
             SenderAccountName = senderAccountName;
         }
+
         public string OutputTransaction()
         {
             string transactionInfo = "Операция: " + this.Operation + Environment.NewLine +
